Renamed package from 'training' to 'ch02'.
