+++ src/main/scala/training/ch02/ex03.scala
@@ -1,4 +1,4 @@
-package training
+package ch02
 
 object ex03 {
