+++ src/main/scala/training/ch02/ex05.scala
@@ -1,4 +1,4 @@
-package training
+package ch02
 
 import ex03._
